Move EDITQUEUE and GLOSSEDITQUEUE to autoloading classes
